chore: deduplicate 5 copy-pasted utility functions
Consolidate identical function bodies that
  VibeDrift's own CodeDNA scanner flagged as exact semantic duplicates:

- shannonEntropy() -> new src/utils/math.ts
  (shared by drift/utils + analyzers/naming)
  - projectHash() -> exported from src/core/baseline.ts (was private in 3 files)
  - isSourceFile() -> removed from 2 drift detectors, use isAnalyzableSource from drift/utils
  - directoryOf() -> removed from return-shape-consistency, imported from drift/utils
  - grade()/getGrade() -> exported from ml-client/coherence.ts, imported by summarize.ts

Self-scan
  codedna-fingerprint findings reduced from 6 to 2. No behavior changes.
